connectors/facebook-marketing/pyproject.toml
[project]
name = "airbyte-agent-facebook-marketing"
version = "0.1.0"
description = "Airbyte Facebook-Marketing Connector for AI platforms"
readme = "README.md"
requires-python = ">=3.13"
license = {text = "Elastic-2.0"}
authors = [
    {name = "Airbyte", email = "contact@airbyte.io"}
]
keywords = ["airbyte", "facebook-marketing", "connector", "api", "ai", "agent", "llm", "mcp", "data-integration"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: Other/Proprietary License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries :: Application Frameworks",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Typing :: Typed",
]
dependencies = [
    "httpx>=0.24.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
    "segment-analytics-python>=2.2.0",
    "jsonref>=1.1.0",
    "jsonpath-ng>=1.6.1",
    "jinja2>=3.0.0",
    "opentelemetry-api>=1.37.0",
    "opentelemetry-sdk>=1.37.0",
]

[project.urls]
Homepage = "https://github.com/airbytehq/airbyte-agent-connectors"
Documentation = "https://docs.airbyte.com/ai-agents/"
Repository = "https://github.com/airbytehq/airbyte-agent-connectors"
Issues = "https://github.com/airbytehq/airbyte-agent-connectors/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["airbyte_agent_facebook_marketing"]

[tool.hatch.build.targets.wheel.shared-data]
"airbyte_agent_facebook_marketing/connector.yaml" = "share/airbyte_agent_facebook_marketing/connector.yaml"